Move Client API Internal stuff to /Internal Move ProjectionsClient and ProjectionsManager to /Projections
