+++ src/EventStore.ClientAPI/EventStore.ClientAPI.csproj
@@ -54,9 +54,12 @@
     <Compile Include="AllEventsSlice.cs" />
     <Compile Include="Common\Log\FileLogger.cs" />
     <Compile Include="ConnectionString.cs" />
-    <Compile Include="Core\GossipSeedEndPointDiscoverer.cs" />
-    <Compile Include="Core\IEventStoreConnectionLogicHandler.cs" />
-    <Compile Include="Core\NodeEndPoints.cs" />
+    <Compile Include="Internal\Consts.cs" />
+    <Compile Include="Internal\Empty.cs" />
+    <Compile Include="Internal\EventStoreNodeConnection.cs" />
+    <Compile Include="Internal\GossipSeedEndPointDiscoverer.cs" />
+    <Compile Include="Internal\IEventStoreConnectionLogicHandler.cs" />
+    <Compile Include="Internal\NodeEndPoints.cs" />
     <Compile Include="GossipSeed.cs" />
     <Compile Include="Internal\IPEndPointExtensions.cs" />
     <Compile Include="StreamCheckpoint.cs" />
@@ -98,32 +101,29 @@
     <Compile Include="DeleteResult.cs" />
     <Compile Include="SystemSettings.cs" />
     <Compile Include="EventReadResult.cs" />
-    <Compile Include="Core\EventStoreConnectionLogicHandler.cs" />
-    <Compile Include="Core\IEndPointDiscoverer.cs" />
-    <Compile Include="Core\OperationsManager.cs" />
-    <Compile Include="Core\StaticEndPointDiscoverer.cs" />
-    <Compile Include="Core\SubscriptionsManager.cs" />
-    <Compile Include="EventStoreNodeConnection.cs" />
+    <Compile Include="Internal\EventStoreConnectionLogicHandler.cs" />
+    <Compile Include="Internal\IEndPointDiscoverer.cs" />
+    <Compile Include="Internal\OperationsManager.cs" />
+    <Compile Include="Internal\StaticEndPointDiscoverer.cs" />
+    <Compile Include="Internal\SubscriptionsManager.cs" />
     <Compile Include="Exceptions\ClusterException.cs" />
     <Compile Include="Exceptions\NotAuthenticatedException.cs" />
     <Compile Include="IEventStoreConnection.cs" />
-    <Compile Include="Core\IEventStoreTransactionConnection.cs" />
+    <Compile Include="Internal\IEventStoreTransactionConnection.cs" />
     <Compile Include="StreamAcl.cs" />
     <Compile Include="SubscriptionDropReason.cs" />
     <Compile Include="Exceptions\AccessDeniedException.cs" />
     <Compile Include="StreamMetadata.cs" />
-    <Compile Include="Core\ClusterDnsEndPointDiscoverer.cs" />
-    <Compile Include="Core\Messages.cs" />
-    <Compile Include="Core\SimpleQueuedHandler.cs" />
-    <Compile Include="Consts.cs" />
+    <Compile Include="Internal\ClusterDnsEndPointDiscoverer.cs" />
+    <Compile Include="Internal\Messages.cs" />
+    <Compile Include="Internal\SimpleQueuedHandler.cs" />
     <Compile Include="EventStoreConnection.cs" />
     <Compile Include="EventStoreCatchUpSubscription.cs" />
-    <Compile Include="Empty.cs" />
     <Compile Include="EventData.cs" />
     <Compile Include="EventStoreSubscription.cs" />
     <Compile Include="Exceptions\EventStoreConnectionException.cs" />
     <Compile Include="Exceptions\ServerErrorException.cs" />
-    <Compile Include="ProjectionsClient.cs" />
+    <Compile Include="Projections\ProjectionsClient.cs" />
     <Compile Include="ReadDirection.cs" />
     <Compile Include="ResolvedEvent.cs" />
     <Compile Include="Exceptions\CannotEstablishConnectionException.cs" />
@@ -147,7 +147,7 @@
     <Compile Include="Messages\ClientMessagesExtensions.cs" />
     <Compile Include="Messages\ClusterMessages.cs" />
     <Compile Include="Position.cs" />
-    <Compile Include="ProjectionsManager.cs" />
+    <Compile Include="Projections\ProjectionsManager.cs" />
     <Compile Include="RecordedEvent.cs" />
     <Compile Include="SliceReadStatus.cs" />
     <Compile Include="StreamPosition.cs" />
