+++ src/EventStore.ClientAPI/RawStreamMetadataResult.cs
@@ -1,4 +1,5 @@
 using EventStore.ClientAPI.Common.Utils;
+using EventStore.ClientAPI.Internal;
 
 namespace EventStore.ClientAPI
 {
